Tidy up pkg/acme package
